+++ R/normal-address.R
@@ -17,14 +17,16 @@
 #'   strings with `NA`.
 #' @return A vector of normalized street addresses.
 #' @examples
-#' normal_address("12 east street, suite 209", abbs = usps_street)
+#' normal_address("12east street, suite209", abbs = usps_street)
 #' @importFrom stringr str_to_upper str_replace_all str_trim str_squish str_replace
 #' @family geographic normalization functions
 #' @export
 normal_address <- function(address, abbs = NULL, na = c("", "NA"), na_rep = FALSE) {
   address2 <- address %>%
     str_normal() %>%
-    stringr::str_replace_all("^P\\sO", "PO")
+    stringr::str_replace_all("^P\\sO", "PO") %>%
+    stringr::str_replace_all("([:digit:]+)([:alpha:]+)", "\\1 \\2") %>%
+    stringr::str_replace_all("([:alpha:]+)([:digit:]+)", "\\1 \\2")
   if (!is.null(abbs)) {
     address2 <- abbrev_full(x = address2, full = abbs)
   }